🥅 chore: fetch 함수 커스텀 훅으로 합치기
